Karen/requirements.txt
discord.py
requests
python-dotenv
google-generativeai
